Reformat Aldermore Bank CSV for FreeAgent
This is a static page which is intended to be used as follows:

1. Download a CSV from your Aldermore Bank account.
2. View this static page.
3. Use the file chooser to select the CSV from your local machine.
4. Click the "Reformat" button.
5. Click the "Download" button.
6. A CSV file suitable for uploading to FreeAgent banking is downloaded.

This uses ruby.wasm [1] to enable the script tags with type "text/ruby"
to run Ruby in the browser.

I've had to use a nightly build [2] rather than the latest release [3]
in order to get CSV parsing to work. I was seeing a SystemStackError
with the latter. This seems closely related to this issue [4] which was
fixed in this PR [5].

Although the nightly build is supposed to support await for promises
inside script tags marked as data-eval="async", this didn't work for me
so I've had to call `then` on the promise passing in a lambda which is a
bit ugly.

[1]: https://github.com/ruby/ruby.wasm
[2]: https://github.com/ruby/ruby.wasm/releases/tag/2023-06-09-a
[3]: https://github.com/ruby/ruby.wasm/releases/tag/2.0.0
[4]: https://github.com/ruby/ruby.wasm/issues/234
[5]: https://github.com/ruby/ruby.wasm/issues/235